Added PlayerClass API
